requirements.txt
langchain>=0.0.267
langchain-community>=0.0.267
python-dotenv>=1.0.0
pypdf>=3.15.0
tiktoken>=0.5.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
langchain-huggingface>=0.0.267